+++ changelog.md
@@ -13,6 +13,7 @@
 * NVA-3705: Configure Android SONAME
 * LLT-3496: Fix is_exit flag when disconnecting from exit node
 * LLT-3499: Fix missing libtelio tag version
+* LLT-897: Enable sending self and member NAT type to moose
 * LLT-3596: Do not fail to start whole meshnet if single peer config is invalid within meshmap
 * LLT-3520: Add missing verification of sideload feature on macOS
 * LLT-3477: Change absolute file path to module path in logger